Fix plugin to be able to update an existing stack
The 'stack' variable was initialized within the try block so when
it was used to determine if the stack existed or not the variable
was out of scope so it always returned 'null'.  I initialized the
variable outside the try block so when it is checked to be null
it will have a value if the stack exists.